pyproject.toml
[tool.poetry]
name = "cuteness"
version = "0.1.0"
description = "A discord bot that provides cute pictures"
authors = ["Natalie Fearnley <nfearnley@gmail.com>"]
license = "MIT"

[tool.poetry.scripts]
cuteness = "cuteness.main:main"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.3.3"
toml = "^0.10.1"
appdirs = "^1.4.4"
aiohttp = "^3.6.2"

[tool.poetry.dev-dependencies]
flake8 = "^3.8.2"
autopep8 = "^1.5.3"
pytest = "^5.4.3"
pytest-asyncio = "^0.12.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
